# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "O-Ushers API",
    "description": "The **O-Ushers** platform connects event organizers with professional ushers (talents).\n\n## Roles\n| Role | Description |\n|------|-------------|\n| **Usher** | A talent who browses events, applies, and manages their profile |\n| **Organizer** | Posts events, manages applicants, marks attendance & reviews |\n| **Organizer Member** | Uses the organizer workspace without event or staff CRUD access |\n| **Organizer Supervisor** | Uses the organizer workspace and manages event operations |\n| **Admin** | Platform-level management: users, events, verification |\n\n## Authentication\nBrowser clients authenticate with secure HTTP-only cookies. API clients may also send the\nshort-lived access token returned by `POST /auth/login` as a Bearer token.\n\n## Pagination\nPaginated endpoints accept `page` and `size` query parameters and return:\n```json\n{ \"data\": [], \"currentPage\": 1, \"totalPages\": 5, \"totalCount\": 42 }\n```\n",
    "version": "1.0.0",
    "contact": {
      "email": "ooushers@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://o-ushers.vercel.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/admin/users/{id}/block": {
      "patch": {
        "tags": [
          "Admins"
        ],
        "summary": "Block a user",
        "description": "Blocks a user account. Business rules:\n- **BR-09**: Admin accounts cannot be blocked.\n- An admin cannot block themselves.\n",
        "operationId": "blockUser",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "User blocked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/admin/users/{id}/unblock": {
      "patch": {
        "tags": [
          "Admins"
        ],
        "summary": "Unblock a user",
        "description": "Unblocks a user account and sets status back to `verified`.",
        "operationId": "unblockUser",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "User unblocked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/admin/users/{id}/verify": {
      "patch": {
        "tags": [
          "Admins"
        ],
        "summary": "Verify a user",
        "description": "Manually sets a user's `isVerified` flag to `true`.",
        "operationId": "verifyUser",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "User verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/admin/users/{id}/unverify": {
      "patch": {
        "tags": [
          "Admins"
        ],
        "summary": "Unverify a user",
        "description": "Sets a user's `isVerified` flag to `false`.",
        "operationId": "unverifyUser",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "User unverified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/admin/users/{id}/status": {
      "patch": {
        "tags": [
          "Admins"
        ],
        "summary": "Update user status (legacy)",
        "description": "Legacy endpoint — sets the user's `status` field directly.\nCannot be used on admin accounts.\n",
        "operationId": "updateUserStatus",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "verified",
                      "blocked"
                    ],
                    "example": "blocked"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User status updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/admin/ushers/{id}/reset-excuses": {
      "post": {
        "tags": [
          "Admins"
        ],
        "summary": "Reset usher late excuses",
        "description": "Resets the `lateExcuseCount` and `consecutiveGoodEvents` counters\nfor a usher account.\n",
        "operationId": "resetExcuses",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Late excuse counter reset",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "lateExcuseCount": {
                              "type": "integer",
                              "example": 0
                            },
                            "consecutiveGoodEvents": {
                              "type": "integer",
                              "example": 0
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/admin/event-action-requests": {
      "get": {
        "tags": [
          "Admins"
        ],
        "summary": "List event action requests",
        "operationId": "getEventActionRequests",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "all"
              ],
              "default": "pending"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Action requests with organizer and event details"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/admin/event-action-requests/{id}": {
      "patch": {
        "tags": [
          "Admins"
        ],
        "summary": "Approve or reject an event action request",
        "operationId": "resolveEventActionRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "decision"
                ],
                "properties": {
                  "decision": {
                    "type": "string",
                    "enum": [
                      "approved",
                      "rejected"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request resolved"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "Request was already resolved"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived JWT obtained from `POST /auth/login`. Browser clients normally\nrely on the secure HTTP-only session cookies instead.\n"
      }
    },
    "parameters": {
...
      "UserIdParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "UUID of the user."
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error or bad input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Validation error: eventId is required"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid Bearer token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Authentication required"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not authorized for this action",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Insufficient permissions"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "User not found"
            }
          }
        }
      },
...
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Something went wrong"
          }
        }
      },
...
    }
  }
}